+++ libft/ft_strlcpy.c
@@ -6,14 +6,14 @@
 /*   By: ede-nada <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/25 14:24:10 by ede-nada          #+#    #+#             */
-/*   Updated: 2020/01/25 14:24:12 by ede-nada         ###   ########.fr       */
+/*   Updated: 2020/01/25 14:38:31 by ede-nada         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned int	ft_strlcpy(char *dest, char *src, unsigned long int size)
+unsigned long int	ft_strlcpy(char *dest, char *src, unsigned long int size)
 {
-	unsigned int	i;
-	unsigned int	j;
+	unsigned long int	i;
+	unsigned long int	j;
 
 	i = 0;
 	while (src[i] != 0)
