+++ libft/ft_strlen.c
@@ -6,11 +6,11 @@
 /*   By: ede-nada <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/21 19:32:13 by ede-nada          #+#    #+#             */
-/*   Updated: 2020/01/21 20:53:23 by ede-nada         ###   ########.fr       */
+/*   Updated: 2020/01/25 14:48:27 by ede-nada         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+unsigned long int	ft_strlen(const char *str)
 {
 	int i;
